engine: fix warnings.

+++ game/engine.cc
@@ -20,7 +20,7 @@ Engine::Engine(Audio& audio, VocalTrackPtrs vocals, Database& database):
 	unsigned i = 0;
 	for (Analyzer& a: analyzers) {
 		// Calculate the space required for pitch frames
-		size_t frames = vocals[i]->endTime / Engine::TIMESTEP;
+		size_t frames = static_cast<size_t>(vocals[i]->endTime / Engine::TIMESTEP);
 		m_database.cur.push_back(Player(*vocals[i], a, frames));
 		++i;
 	}
